modularize signal ops
