+++ tfjs-core/src/tests.ts
@@ -87,12 +87,15 @@ import './ops/equal_test';
 import './ops/expand_dims_test';
 import './ops/eye_test';
 import './ops/floor_test';
+import './ops/frame_test';
 import './ops/fused_test';
 import './ops/gather_nd_test';
 import './ops/gather_test';
 import './ops/gram_schmidt_test';
 import './ops/greater_equal_test';
 import './ops/greater_test';
+import './ops/hamming_window_test';
+import './ops/hann_window_test';
 import './ops/hinge_loss_test';
 import './ops/huber_loss_test';
 import './ops/in_top_k_test';
@@ -146,7 +149,6 @@ import './ops/scatter_nd_test';
 import './ops/selu_test';
 import './ops/sigmoid_cross_entropy_test';
 import './ops/sign_test';
-import './ops/signal_ops_test';
 import './ops/slice_test';
 import './ops/slice_util_test';
 import './ops/softmax_cross_entropy_test';
@@ -155,6 +157,7 @@ import './ops/space_to_batch_nd_test';
 import './ops/sparse_to_dense_test';
 import './ops/spectral_ops_test';
 import './ops/stack_test';
+import './ops/stft_test';
 import './ops/strided_slice_test';
 import './ops/sub_test';
 import './ops/sum_test';
